added diabetes data
Steps to reproduce

they are taken 1000 patients and cover three classes (Diabetic, Non-Diabetic, and Predicted- Diabetic). Diabetes is an opportune disease which has large wealth of data available and has with it huge complications. There is a need for a better and a more accurate approach in the diagnosis of the disease

Institutions

University of Information Technology

Categories

Applied Sciences, Health Sciences

Licence

CC BY 4.0